+++ src/main/java/top/longmarch/douyin/controller/DouYinHotsearchController.java
@@ -25,7 +25,7 @@ public class DouYinHotsearchController {
 
     @GetMapping("/hotsearchTrendingSentencesGet")
     public Result hotsearchTrendingSentencesGet(@RequestParam(required = false, defaultValue = DouyinParam.COUNT) Integer count,
-                                                @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Integer cursor) {
+                                                @RequestParam(required = false, defaultValue = DouyinParam.CURSOR) Long cursor) {
         return Result.ok().add(douYinHotsearchService.hotsearchTrendingSentencesGet(count, cursor));
     }
