+++ src/main/java/top/longmarch/douyin/service/DouYinHotsearchService.java
@@ -24,7 +24,7 @@ public interface DouYinHotsearchService {
      * @param cursor 分页游标, 第一页请求cursor是0, response中会返回下一页请求用到的cursor, 同时response还会返回has_more来表明是否有更多的数据。
      * @return
      */
-    HotsearchHotsearchInlineResponse2001Data hotsearchTrendingSentencesGet(Integer count, Integer cursor);
+    HotsearchHotsearchInlineResponse2001Data hotsearchTrendingSentencesGet(Integer count, Long cursor);
 
     /**
      * 获取热点词聚合的视频
